@@ -1,9 +1,8 @@
 {
   "name": "ts-lite-store",
-  "version": "1.0.8",
+  "version": "1.0.36",
   "scripts": {
-    "prepublishOnly": "npm run compile",
-    "compile": "npm run clean && tsc -p .",
+    "build": "npm run clean && npm run copy && tsc -p .",
     "clean": "del /Q /S lib",
     "copy": "xcopy package.json lib\\package.json",
     "test": "mocha --require ts-node/register test/**/*.ts"
@@ -26,11 +25,11 @@
     "chai": "^4.2.0",
     "mocha": "^6.0.0",
     "ts-node": "^8.0.2",
-    "typescript": "^3.3.1"
+    "typescript": "^4.5.4"
   },
   "dependencies": {
-    "@types/react": "16.4.16",
-    "react": "16.5.2",
-    "react-dom": "16.5.2"
+    "@types/react": "^17.0.37",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
   }
 }
